+++ apps/web/lib/inngest/index.ts
@@ -1,3 +1,0 @@
-export { inngest } from "./client";
-export { cleanupProjectSandbox } from "./functions/cleanup-project-sandbox";
-export { startSandbox, stopSandbox } from "./functions/session-sandbox";